Add functionality to initialize the databse in the first application launch
