+++ package.json
@@ -43,7 +43,8 @@
     "react-native-svg": "^15.4.0",
     "react-native-toast-message": "^2.2.0",
     "react-native-uuid": "^2.0.2",
-    "react-native-web": "~0.19.10"
+    "react-native-web": "~0.19.10",
+    "@react-native-async-storage/async-storage": "1.23.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
