+++ services/RutinasServices.ts
@@ -207,67 +207,3 @@ export async function saveAndFinishRutinaCabeceraAndDetalle(detalle: RutinaDetal
     await db.insert(RutinaDetalle).values(filteredDetalles)
     await db.update(RutinaCabecera).set({finished: true}).where(eq(RutinaCabecera.id, detalle[0].idRutinaCabecera))
 }
-
-const crearCabecera = async (rutinaId)=>{
-    await db.insert(RutinaCabecera).values({
-        id: 2,
-        idRutina: rutinaId,
-        fecha: new Date().toISOString()
-    })
-    await db.insert(RutinaDetalle).values([
-        {idRutinaCabecera: 1,
-        idEjercicio: 1,
-        peso: 50,
-        reps: 25,
-        numeroSerie: 1,},
-        {idRutinaCabecera: 1,
-        idEjercicio: 2,
-        peso: 25,
-        reps: 25,
-        numeroSerie: 1,}
-]
-    )
-}
-
-export async function initializeDB(){
-    await initializeEjercicios()
-    await initializeRutinas()
-}
-
-async function initializeEjercicios(){
-    await db.insert(Ejercicio).values([
-        {
-            id: 1,
-            name: "Press Banca"
-        },
-        {
-            id: 2,
-            name: "Press Militar"
-        },
-        {
-            id: 3,
-            name: "Sentadillas"
-        }
-    ])
-} 
-
-async function initializeRutinas(){
-    
-    await db.insert(Rutina).values([
-        {
-            id: 1,
-            name: "Torso"
-        }
-    ])
-    await db.insert(RutinaToEjercicio).values([
-        {
-            rutinaId: 1,
-            ejercicioId: 1
-        },
-        {
-            rutinaId: 1,
-            ejercicioId: 2
-        }
-    ])
-}
-//export function saveTrain()
